Creation of all pages for the application
All pages for the application was created, the entry and landing
page which is the index.html, page where a user can view all books in the
library which is shelf.html, page to borrow - borrow.html, page to access profile and see books yet to be returned, page to borrow books and
the page with yet to be returned books below and a page for the user to see the borrowing history.

custom css file (main.css) was added for custom styling, for default styling Materialize CSS Framework is being used. Also,
custom js scripts file was added for custom scripts.

HTML, CSS Design for all pages hence implemented.