Controller for user actions.
A method "User.create" that signs up user by creating their records in the database, the user's password is hashed, in the migration definition, the email and username fields are made unique. if all goes well record is created else. Error is issued.

Also another method "User.findOne" that checks if a has been signed up, if all goes well user gets signed in and is issued a token, else error is issued.